% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkcd.R
\name{xkcd}
\alias{xkcd}
\alias{rxkcd}
\alias{dxkcd}
\alias{pxkcd}
\alias{qxkcd}
\title{The xkcd Distribution}
\source{
\code{rxkcd} is based on \href{https://xkcd.com/2118/}{xkcd comic 2118},
and a more detailed version (including pdf and cdf) is given by Charles Geyer
in \href{http://www.stat.umn.edu/geyer/8054/hw/}{Doing Assignment 2},
though the implementation here is a little bit different (simplified some
expressions).
}
\usage{
rxkcd(n, nsd = 1, swap.end.points = FALSE, flag = "R")

dxkcd(y, nsd = 1, log = FALSE, swap.end.points = FALSE, flag = "R")

pxkcd(y, nsd = 1, log.p = FALSE, swap.end.points = FALSE, flag = "R")

qxkcd(
  p,
  nsd = 1,
  log.p = FALSE,
  swap.end.points = FALSE,
  tol = .Machine$double.eps,
  flag = "R"
)
}
\arguments{
\item{n}{Number of observations. If \code{length(n) > 1}, the length is taken
to be the nubmer required and a warning message will be produced.}

\item{nsd}{Vector of standard deviations for underlying normal distribution.}

\item{swap.end.points}{Logical; if TRUE, \code{rxkcd, dxkcd, pxkcd, qxkcd}
are for distribution of \eqn{1/[\sqrt{2\pi}\sigma] - Y} rather than
\eqn{Y} itself.}

\item{flag}{Indicator of using R or C for performing core functions. It can
only take value from \code{c("R", "C")}.}

\item{y}{Vector of quantiles.}

\item{log, log.p}{Logical; if TRUE, \code{dxkcd} and \code{pxkcd} will give
densities and probabilities in log scale; \code{qxkcd} will regard its first
argument as log probabilities.}

\item{p}{Vector of probabilities.}

\item{tol}{The desired accuracy for \code{qxkcd} when numerically inversing
\code{pxkcd}}
}
\value{
\code{dxkcd} gives the density,
\code{pxkcd} gives the distribution function,
\code{qxkcd} gives the quantile function, and
\code{rxkcd} generates random numbers for xkcd distribution.

Logical arguments \code{log}, \code{log.p} and \code{swap.end.points}
can't accept a logical vector of length greater than 1. Other than that, all
the other arugments (first argument of each function and \code{nsd}) will
broadcast according to each other.

\code{nsd = 0} will produce error since it makes no sense to generate a
nonnegative number that is uniformly distributed on \eqn{(0, Inf)}.

\code{qxkcd} is calculated by numerically inversing \code{pxkcd} via
\code{stats::uniroot}. The calculation accuracy can be set up via \code{tol}
argument. The default accuracy is \code{.Machine$double.eps}.
}
\description{
Density, distribution function, quantile function and random number
generation for the xkcd distribution with the underlying normal standard
deviation equal to \code{nsd}.
}
\details{
The default value of some arguments can be seen in the Usage section
above.

The xkcd distribution is derived from a normal distribution
\eqn{X ~ N(0, \sigma^2)} (with density \eqn{f(x)}). Given \eqn{X=x}, we
generate \eqn{Y|X=x ~ Uniform(0, f(x))} accordingly. The resulting marginal
distribution of \eqn{Y} is called xkcd distribution and we denote it as
\eqn{Y ~ xkcd(\sigma)}.

\eqn{Y} is in the range \eqn{(0, 1/[\sqrt{2\pi}\sigma])}, and it has
distribution function
\deqn{2P[X\le -h(y)]+2yh(y), Y is in (0, 1/[\sqrt{2\pi}\sigma])}
and density function
\deqn{2h(y), Y is in (0, 1/[\sqrt{2\pi}\sigma])}
where \eqn{h(y)} is the nonnegative root of \eqn{f(x)=y}.
}
\examples{
# some tests, see formal tests in unit test
testy <- dnorm(0)/2
dnorm(dxkcd(testy)/2) == testy
dxkcd(0) == Inf
dxkcd(dnorm(0)) == 0

# issue of numerically inversing pxkcd()
pxkcd(qxkcd(0.5)) == 0.5

# check the density
Y <- rxkcd(1000)
W <- rxkcd(1000, swap.end.points = TRUE)
dxkcd_w <- function(w) dxkcd(w, swap.end.points = TRUE)
par(mfrow = c(1, 2))
hist(Y, freq = FALSE, breaks = 20)
curve(dxkcd, 0, dnorm(0), add = TRUE, lwd = 2)
hist(W, freq = FALSE, breaks = 20)
curve(dxkcd_w, 0, dnorm(0), add = TRUE, lwd = 2)
par(mfrow = c(1, 1))

}
\references{
xkcd comic 2118, \emph{How To Annoy A Statistician},
\url{https://xkcd.com/2118/}

Charles Geyer, \emph{Doing Assignment 2},
\url{http://www.stat.umn.edu/geyer/8054/hw/}
}
\seealso{
\href{https://stat.ethz.ch/R-manual/R-devel/library/stats/html/Distributions.html}{Distributions} for other standard distributions.
}
